Introduce macro definitions.
A macro have to be defined before it is used, otherwise
it is treated as function call.
For example
	f(x) := sin(x)	# Macro. Note `:=' operator.
	x' = f(t+1)
is interpreted as
	x' = sin(t+1)
without the need to call `f' when expression evaluated.

At the moment macro is restricted to have 0 or 1 argument,
because funtion and macro calls are sintactically undistinguishable
and arglist format used for function call is not appropriate for macro
call.